requirements.txt
torch>=1.13.1
torchvision>=0.14.1
numpy>=1.21.6
matplotlib>=3.5.3
Pillow>=9.4.0
argparse>=1.4.0
tqdm>=4.65.0  # for progress bars
